+++ ffi/diplomat/cpp/docs/source/properties_sets_ffi.rst
@@ -1,5 +1,5 @@
-``properties::ffi``
-===================
+``properties_sets::ffi``
+========================
 
 .. cpp:class:: ICU4XUnicodeSetProperty
