+++ ffi/diplomat/src/lib.rs
@@ -38,7 +38,7 @@ pub mod fixed_decimal;
 pub mod locale;
 pub mod locale_canonicalizer;
 pub mod pluralrules;
-pub mod properties;
+pub mod properties_sets;
 pub mod provider;
 
 #[cfg(target_arch = "wasm32")]
