+++ ffi/diplomat/src/properties_sets.rs
